requirements.txt
# Core
playwright>=1.40.0
apscheduler>=3.10.4
aiohttp>=3.9.0
pyyaml>=6.0.1

# Cookie extraction from Chrome
pycookiecheat>=0.7.0
rookiepy>=0.5.0

# MCP server
mcp>=1.0.0

# Notifications
python-telegram-bot>=20.0
plyer>=2.1.0

# Optional: better async handling
colorlog>=6.8.0
